## Description
This PR implements a comprehensive UI overhaul focused on "sleek"
aesthetics, precise alignment, and improved stability for the `Rescan`
functionality. It addresses multiple visual inconsistencies in the
FileList, Settings, and HelpModal components, ensuring pixel-perfect
column alignment and robust rendering of blank line spacers in terminal
environments.

## Key Changes

### 🎨 Visual & Layout
- **FileList Alignment:** 
- Implemented strict column gaps (1-char width) in the layout
calculation, headers, and data rows.
- Aligned "Size", "Usage", and "Percent" headers perfectly with their
data columns.
- Added a visual gap between the percentage text and the usage bar `[`.
- **Modals (Settings & Help):**
- **Sticky Headers:** Implemented sticky header logic for better
navigation in long lists.
- **Spacers:** Forced rendering of blank lines between sections by using
`<Box><Text> </Text></Box>` to prevent terminal collapsing.
- **Help Menu:** Enforced strict item ordering (e.g., ensuring "Rescan"
appears correctly) and fixed layout artifacts where longer text would
overwrite shorter text during scrolling.

### ⚙️ Functionality & Stability
- **Rescan Feature:** 
  - Added `RESCAN` action (Key: `R`) to keys.ts and App.tsx.
- **OOM Prevention:** Implemented a safety delay (50ms) during the
rescan process to allow the Garbage Collector (GC) to reclaim memory
from the dropped file tree before allocating the new one, preventing
Out-Of-Memory crashes on large directories.
- **Context-Aware Footer:** Updated the footer to display dynamic key
hints based on the active mode (Default vs. Settings vs. Help).

### 🧹 Refactoring
- Refactored HelpModal and Settings to share consistent row-mapping
logic for selection and scrolling.
- Updated `useScrollableList` to handle modal navigation more reliably.

## Verification
- [x] **FileList:** Columns align perfectly; divider line has gaps
matching the columns.
- [x] **Rescan:** Pressing `R` correctly releases memory and scans again
without crashing.
- [x] **Settings/Help:** Blank lines appear between sections; Headers
stick to the top during scroll.
- [x] **Walkthrough:** Validated against the updated `walkthrough.md`
checklist.

+++ src/components/ConfirmDelete.tsx
@@ -10,7 +10,7 @@ interface ConfirmDeleteProps {
 export const ConfirmDelete: React.FC<ConfirmDeleteProps> = ({ fileName, theme }) => {
   return (
     <Box
-      borderStyle="double"
+      borderStyle="single"
       borderColor="red"
       flexDirection="column"
       padding={1}
